types, evalOrder, lists and refactoring
Added an enum based type system for the elements of the list. Changed the way the list is evaluated so that function prefixing is no longer required. started to work on lists that return themselves, in normal List that would be '(a b c). I want to do it like [a b c]. Did some general refactoring.